refactor(mobile): layout compartilhado das telas
