+++ mobile/app/memories.tsx
@@ -1,9 +1,9 @@
 import { Text, View } from 'react-native'
 
-export function Memories() {
+export default function Memories() {
   return (
     <View className="flex-1 items-center justify-center">
-      <Text>Memories</Text>
+      <Text className="text-lg text-gray-500">Memories</Text>
     </View>
   )
 }
